+++ webhook_logger.go
@@ -1,43 +1,9 @@
 package ayame
 
 import (
-	"fmt"
-	"io"
-	"os"
-	"time"
-
 	"github.com/rs/zerolog"
-	"gopkg.in/natefinch/lumberjack.v2"
 )
 
 func InitWebhookLogger(config *Config) (*zerolog.Logger, error) {
-
-	if f, err := os.Stat(config.LogDir); os.IsNotExist(err) || !f.IsDir() {
-		return nil, err
-	}
-
-	logPath := fmt.Sprintf("%s/%s", config.LogDir, config.WebhookLogName)
-
-	// https://github.com/rs/zerolog/issues/77
-	zerolog.TimestampFunc = func() time.Time {
-		return time.Now().UTC()
-	}
-
-	writer := &lumberjack.Logger{
-		Filename:   logPath,
-		MaxSize:    config.LogRotateMaxSize,
-		MaxBackups: config.LogRotateMaxBackups,
-		MaxAge:     config.LogRotateMaxAge,
-		Compress:   config.LogRotateCompress,
-	}
-
-	writers := io.MultiWriter(writer)
-	// デバッグが有効な時はコンソールにもだす
-	if config.Debug {
-		writers = io.MultiWriter(writers, writer)
-	}
-
-	logger := zerolog.New(writers).With().Timestamp().Logger()
-
-	return &logger, nil
+	return InitLogger(config, config.WebhookLogName, "webhook")
 }