+++ cmd/ayame/main.go
@@ -31,11 +31,18 @@ func main() {
 		log.Fatal(err)
 	}
 
-	// グローバルの logger に代入する
-	logger, err := ayame.InitLogger(config, config.LogName, "ayame")
+	// グローバルの logger を初期化
+	err = ayame.InitLogger(config)
 	if err != nil {
 		log.Fatal(err)
 	}
+
+	// ayame.jsonl ロガーを用意
+	logger, err := ayame.NewLogger(config, config.LogName, "ayame")
+	if err != nil {
+		log.Fatal(err)
+	}
+	// ayame.jsonl は グローバルの logger に代入する
 	zlog.Logger = *logger
 
 	config.PrintConfig()